Extract URL normalizers into src/types/url-normalize.ts
Break a runtime import cycle between src/types/argument-roles.ts and
src/trusted-process/domain-utils.ts. The ArgumentRole registry needs
canonicalize() functions for fetch-url and git-remote-url roles, but
those lived in trusted-process/, which in turn imports ArgumentRole as
a type — a runtime cycle through types/, which is supposed to be a
sink layer.

Move the two pure URL canonicalizers (normalizeUrl, normalizeGitUrl)
into a new neutral module src/types/url-normalize.ts. Domain matching,
IP detection, domain extraction, and git remote resolution stay in
trusted-process/ since they're consumed by the policy engine.

Pure refactor — no behavior change, no signature change.

Addresses an item from the recent layer-violation audit.
